requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
sqlalchemy[asyncio] >= 2.0
pylint>=3.0
psycopg2>=2.8.6
pynvim==0.5
asyncpg >=0.29
python-dotenv>=1.0
strawberry-graphql[fastapi] == 0.219.1
strawberry-sqlalchemy-mapper == 0.4.1
